requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
sqlalchemy-stubs==0.4